requirements-dev.txt
pytest>=8.0
pytest-cov>=5.0
pytest-django>=4.8
pre-commit>=3.7
ruff>=0.5
